Use stable names for hook impls.
We were using memory addresses of functions as part of their name when
creating hook implementations. These names were not stable across
invocations which made the output of e.g., `spicyc` not reproducible.
This e.g., means that the output of `spicyc` was not reproducible, and
could lead to different names for the same function when generating code
and linker glue separately.

With this patch we use a hash of the function's source location instead
of its memory address instead. This means that we get a deterministic
name which only depends on the input file.