Require all callbacks executing on fibers to return something.
We previously were conditionally synthesising a boolean return value
when executing a `void` callback on a fiber while keeping the original
return value for non-`void` callbacks. This required us to create a
wrapping lambda in a header which caused the instantiation of additional
`std::function` specializations.

By requiring all callbacks executing on fibers to be non-`void` we can
remove the wrapper and cut down the number of additional `std::function`
specializations we need to instantiate. Additionally, this also cleans
up semantics as we do not need to synthesise artifical return values
which might clash with user expectations anymore.

During JIT of a typical parser this saves about 250ms in compiler
frontend time.